pyproject.toml
[tool.poetry]
name = "bluebash-test1"
version = "0.1.0"
description = ""
authors = ["narendra-bluebash <narendra.dangi@bluebash.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.37"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
colorlog = "^6.9.0"
python-multipart = "^0.0.20"
twilio = "^9.4.2"
alembic = "^1.14.0"
livekit = "^0.20.1"
livekit-api = "^0.8.2"
psycopg2 = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
